Added property creation class methods
Properties are objects containing information about the type and the
value. The value can be a key, a string, a double, float, or an
integer. If the value is a double or float, the property can also
contain errors (both positive or negative) and units (defined by a
key). No unit keys are defined at the moment.