Save properties file
Saving of document metadata in file bundle. Data include original file
name, file name in bundle, data object type, and properties.
DataObjectWrapper can be translated into a serializable dictionary,
which is saved in the bundle. The data (and thumbnail) for each data
object is not yet saved into the bundle. The data object needs to be
saved (and read) incrementally.