BUGFIX: Extra hack for saving duplicates of Unsaved documents
The duplicate copy did not work when the original document had never
been saved ('Untitled'). In that case, the data object in the original
file (which should be present as the file has not been incrementally
loaded yet) is serialized into binary data and that data is used to
create a copy of the data object which is then added to the duplicate
document.